share_learning/requirements.txt
Django              4.1.1
mysqlclient         2.1.1
djangorestframework 3.13.1
drf-nested-routers   0.93.4
django-debug-toolbar 3.6.0